healthcare_app/pyproject.toml
[project]
name = "healthcare-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr Balashenko",email = "balashenkoaleksandr294@gmail.com"}
]
package-mode = false
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.11,<6.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "ruff (>=0.15.2,<0.16.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "dj-database-url (>=3.1.2,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "isort (>=8.0.0,<9.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
]

[tool.black]
line-length = 120
force-exclude = "/(migrations)/"


[tool.ruff]
fix = false
lint.select = ["I001", "E402", "Q000", "ANN001", "ANN002", "ANN003", "ANN201", "ANN202", "ANN204", "ANN205", "ANN206", "F401", "F541", "E712", "W291", "E302", "A003", "A002", "A001"]
lint.ignore = ["ANN002", "ANN003", 'ANN204']
lint.fixable = []
lint.unfixable = []
line-length = 120
exclude = ["migrations", "venv", ".venv", ".env", "manage.py"]

[tool.ruff.lint.per-file-ignores]
"*/filters.py" = ["ANN001", "ANN201", "ANN202"]
"*/views.py" = ["ANN001", "ANN201", "ANN202"]
"*/permissions.py" = ["ANN001", "ANN201", "ANN202"]
"*/mixins.py" = ["ANN001", "ANN201"]
"*/*serializers*.py" = ["ANN001", "ANN201", "ANN204"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
